+++ src/__init__.py
@@ -1,6 +1,6 @@
 """Public package interface for CoolBox."""
 
-__version__ = "1.0.82"
+__version__ = "1.0.83"
 
 import os